refactor: 역할 확장을 위한 타입 구조 개선 - Teacher 타입에 접두사 추가 - features/mypage/types에 UserBasicInfo union 타입 추가 - 공통 컴포넌트는 UserBasicInfo 사용하도록 변경
의존성 방향 유지를 위해 shared/components/profile 제거 -> feature/profile/components로 이동
- features -> entities -> shared 의존성 방향 유지
- shared -> features 방지
- profile-card는 profile 기본적으로 컴포넌트 + mypage에서 관리할 때만 사용하므로 범용 UI가 아니라고 판단  (비즈니스 도메인 컴포넌트)
