+++ src/entities/teacher/types/index.ts
@@ -2,8 +2,10 @@ import { domain } from '@/entities/teacher/core/teacher.domain';
 import { dto, payload } from '@/entities/teacher/infrastructure/teacher.dto';
 import { z } from 'zod';
 
-export type BasicInfoDTO = z.infer<typeof dto.basicInfo>;
+export type TeacherBasicInfoDTO = z.infer<typeof dto.basicInfo>;
 
-export type FrontendBasicInfo = z.infer<typeof domain.basicInfo>;
+export type FrontendTeacherBasicInfo = z.infer<typeof domain.basicInfo>;
 
-export type UpdateBasicInfoPayload = z.infer<typeof payload.updateBasicInfo>;
+export type UpdateTeacherBasicInfoPayload = z.infer<
+  typeof payload.updateBasicInfo
+>;
