+++ src/features/mypage/types/index.ts
@@ -1,0 +1,9 @@
+import {
+  FrontendTeacherBasicInfo,
+  UpdateTeacherBasicInfoPayload,
+} from '@/entities/teacher';
+
+// TODO Student, Parent 추가
+export type UserBasicInfo = FrontendTeacherBasicInfo;
+
+export type UpdateUserBasicInfoPayload = UpdateTeacherBasicInfoPayload;
